Moved Client related classes in seperate package.
