+++ ContextApp/src/main/java/com/github/migi_1/ContextApp/client/ClientWrapper.java
@@ -1,4 +1,4 @@
-package com.github.migi_1.ContextApp;
+package com.github.migi_1.ContextApp.client;
 
 import com.github.migi_1.ContextMessages.AccelerometerMessage;
 import java.io.IOException;
